fix(aws): Fallback if asg instances cannot be described (#3383)
There has been some back and forth around the method used to determine
instances to enable/disable.

This PR introduces a fallback lookup based on instances tagged with the
asg name in the event that the original lookup failed.

Historically we looked images up this way but moved away due to slowness
in larger account/regions. It seems reasonable to at least try it as a
fallback rather than letting the operation fail.