fix(artifacts): Fix threading bug in HTTP artifact credentials (#3390)
* test(artifacts): Add tests to ArtifactCredentials classes

This also fixes a bug in the BitbucketArtifactCredentials that
was caught by the new unit test.

* refactor(artifacts): Optimize imports

Optimize imports on the module to make upcoming diffs cleaner

* refactor(artifacts): Move copy-pasted exception to its own class

The exception FailedDownloadException is duplicated in 4 different
classes; pull it out into its own file.

* refactor(artifacts): Fix some code style issues

- Remove some unused methods
- Remove @Data from the Credentials classes as we don't want the full
  external state visible.  Instead, just put @Getter on the fields we
  want exposed.
- Explicitly add imports that were transitively included
- Set some fields as final
- A few minor style fixes

* fix(artifacts): Fix threading bug in HTTP artifact credentials

Many artifact credentials duplicate the same logic of:
- Getting an authorization header from the artifact account config
- Generating a URL from an artifact, and sending a request to download it

This logic is not thread-safe, which was recently fixed in the Github
artifact handler.  This commit pulls the common logic into a base class
and removes the duplicate (and non-threadsafe) implementations.