test(redis): Use embedded redis for unit tests (#463)
The unit tests for front50-redis depend on there being an existing
redis listening on localhost:6379.  If so, they send data to that
redis (which may not be intended to receive test data); if not, the
tests are silently skipped. Instead, set up a test embedded redis
and use that.