fix(redis): Fix leaked embedded redis from tests (#335)
Embedded redis is currently defined as a prototype bean in the
test config; this means that a new embedded redis will be generated
any time a redis is need in the auotwiring config. We're currently
only destroying one of them (the one directly asked for in the
test class). Instead, make this a singleton bean with a spring-managed
destruction method, and use the same bean everywhere.