fix(artifacts): Make artifact resolution idempotent (#2731)
Artifact resolution is not currently idempotent; this causes
occasional bugs where trying to resolve again causes errors due
to duplicate artifacts. In order to be truly idempotent, also
changed the Set instances in the resolution to LinkedHashSet
so that the order of operations (and resulting artifact lists
in the pipeline) are stable and deterministic.