refactor(triggers): Gate calls to igor from orca (#2748)
Now that echo handles augmenting triggers with build info, and that
manual triggers default to go through echo, all triggers should be
arriving in Orca with their build information already populated. We
should gate the logic in Orca to only run if it's not there.

We can't completely remove this logic yet because while manual
triggering via echo defaults to enabled, there's still a flag to
turn it off. Once that flag is deprecated, and we're confident that
all manual triggers (including any via the API) go through echo,
we can completely remove this block of code.

This commit also completely removes the populating of taggedImages,
which has not been used since spinnaker/orca#837.