feat(gcb): Add endpoint for listing accounts (#412)
The UI for the GCB stage should provide a drop-down that lists
available accounts, so we need to support this operation on the
back-end.

This is similar to the '/masters' endpoint that lists the available
build services for other services. I didn't have GCB implement
BuildService because the types of operations it supports are different,
which means that we don't get the '/masters' functionality for free.

As I develop this functionality further I'll see whether there's a more
general interface that emerges that this and other build services can
implement, at which point we may be able to add these into the general
'/masters/ endpoint.  But for now, adding a separate '/gcb/accounts'
endpoint that will list GCB accounts.