feat(kubernetes): Use streaming job executor for k8s manifests (#3531)
* feat(core): Add streaming Job functionality

Currently the LocalJobExecutor accumulates all output from the job
into a stream and returns the entire output (as a string) to the
calling function.  In some cases, it would be useful to allow the
calling function to process the output as it is produced to reduce
the amount of output that needs to be stored (and to reduce the
number of times that it is copied around).

Add an optional parameter to runJob specifying a function that will
consume the output as it is produced. When this parameter is passed,
all output will be passed to the supplied callback, and the output
on the JobResult will be null.

* feat(kubernetes): Use streaming job executor for k8s manifests

When fetching lists of kubernetes manifests, the returned
data can be quite large for large clusters. Reduce the amount
of times this data needs to be copied by using a streaming
JSON parser to parse the output as it is produced.

* refactor(core): Improve interface of JobExecutor

* JobStatus is now parametrized and can return any type in its
  stdOut field.  (This field will be renamed output in a follow-up
  commit.)
* StreamConsumer is now parametrized and returns the results of
  its processing on the output stream as a type T instead of
  mutating an array in the calling method.
* Pushed a flag indicating if the job was killed up to JobStatus,
  which eliminated the need for the (almost identical) ExecuteResult
* Changed some of the control flow in JobExecutorLocal.
* Explicitly stated that implementations of StreamConsumer must close
  the stream, and updated the existing implemenation to do so.

* refactor(core): Improve interrupt handling

If the thread is interrupted, we should kill the forked job that is
running, and set the interrupt flag on the thread before returning.

* refactor(core): Rename JobStatus and member variables

JobStatus is really holding the result of a job; it is never
generated until the job is finished.  Rename it to JobResult.

Now that stdOut can be either stdOut or some processing of the
output, rename it to the more general output.  Also rename
stdErr as error, though either would probably be appropriate for it.

* refactor(core): Change StreamConsumer to ReaderConsumer

We'll likely only ever want to read characters from a job, so
push the conversion from InputStream -> InputStreamReader to the
JobExecutor.

Also:
* Don't explicitly specify a character class; just use the default
* buffer the input stream before passing it to the ReaderConsumer

* refactor(core): A few minor changes

* Actually close the JsonReader (and thus the underlying reader/stream),
  which I thought I had done in an earlier commit.
* Rename ResultSupplier to RequestExecutor which is a bit more
  descriptive.