chore(migrations): ignore empty migrations (#2846)
Currently, any non-standard migrations are defined in the orca.yml as a list.
Sometimes, it's nice to be able to run without those migrations
(e.g. when developing both in OSS and private land). However, due to an issue in
spring (which, I believe, should be fixed in boot2.0.3) it's impossible to
override a list with an empty list, but you can override a list with a new list.
Hence this change to not run empy entries