fix(triggers): prevent double-triggering for pipelines with >1 triggers (#529)
* fix(triggers): prevent double-triggering for pipelines with >1 triggers

A pipeline with 2 triggers and an event matching one of them could
actually be triggered twice. The problem is that withMatchingTrigger
would consider the event and all triggers of the pipeline, and it would
be called once for each trigger.

The key fix is to matchTriggerFor(event) in getMatchingPipelines().
As a result, we can change the inputs of getMatchingPipelines to an event
and a matching trigger instead of an event and a pipeline (with all its
triggers).