feat(gcb): Add endpoints for storing and retrieving build status (#425)
* fix(gcb): Fix deserialization of Build object

Jackson's ObjectMapper does not properly deserialize Build objects,
as nested properties that extend com.google.api.client.json.GenericJson
(which itself implements Map) are are deserialized as Maps rather than
as the specific type.

There is probably a way of configuring ObjectMapper to do this properly
but I have not (yet) been able to find it. As a workaround, use the
Google Cloud supported deserialization classes to deserialize this
JSON. In order to keep this workaround localized, make a small
class that will handle deserializing Build classes.

This commit also adds BuildOptions to the build in the integration
test; adding this causes the test to fail when using the standard
ObjectMapper for deserialization.

* feat(gcb): Add endpoints for storing and retrieving build status

Create and endpoint to accept the current status of a build and store
it in Redis.  This endpoint is intended to be called by echo when it
receives a PubSub message about a build status change.  We take care
to look at the timestamp of incoming messages and only update the
cache if the incoming information is newer, as PubSub does not guarantee
delivery order.

Create an endpoint to return the current status of the build, as
per the cache.  This commit does not implement any fallback to
calling the API if the cache does not contain the build; this
will be added in a future commit.

* feat(gcb): Create build should return only the build

When creating a build, we should return only the build instead of the
operation. In general, callers will only be interested in the id of the
created build, and it will be easier for callers to just look up the
"id" of the returned object instead of looking at the deeply nested
operation -> metadata -> build -> id field.