refactor(provider/kubernetes): Remove reflection from cache constructors (#3597)
* refactor(kubernetes): Runtime -> compile time error

All subclasses of KubernetesV2Caching agent need to implement
either primaryKinds or primaryKind, and a failure to do so is
a runtime error.  Make this a compile-time error by standardizing
on primaryKinds and making it abstract in the base class.

* refactor(provider/kubernetes): Remove reflection from cache constructors

Constructing caching agents currently uses reflection, which breaks
static analysis and makes it harder to reason about and refactor this
code. Instead of having each handler return a reference to the class
that should be constructed, instead have the handler return a functional
interface to the actual constructor so that the base class does not need
to use reflection.