refactor(provider/kubernetes): Add tests and simplify cache refresh (#2869)
* test(provider/kubernetes): Add tests to ManifestForceCacheRefresh

Most of the functionality in ManifestForceCacheRefreshTask is not
tested. Add significant test coverage to this class to prepare for
a bug fix/refactor.

* refactor(provider/kubernetes): Replace nested maps with an object

The ManifestForceCacheRefreshTask currently keeps track of its manifests
as a Map<String, List<String>> which leads to a lot of complex iteration
over this structure. Create a new class ScopedManifest and flatten this
structure into a List<ScopedManifest> so that it's much easier to
follow the processing that the class is doing.

* refactor(provider/kubernetes): Add account to ScopedManifest

Rather than thread account through multiple calls in this class, add
it to the ScopedManifest class and set the account on each manifest
when we create the initial list.

This also makes the helper class Details identical to ScopedManifest,
so replace instances of Details with ScopedManifest.

Finally, replace pendingRefreshProcessed which both returns a status
and mutates the stage context with getRefreshStatus, which only
returns a status. Leave mutation of the context to the caller,
checkPendingRefreshes. This allows us to avoid needing to mutate
a ScopedManifest and keep the class immutable.

* refactor(provider/kubernetes): Track ScopedManifests directly

Now that we have a simple data class to represent a manifest to
refresh that implements equals and hashCode, we don't need to
manually serialize it with toManifestIdentifier. Just directly
add these manifests to the collections we're using to track
them.