feat(gcb): Monitor GCB build status after starting a build (#2875)
* refactor(gcb): Use a GoogleCloudBuild type as return from igor

To make the task logic simpler, create a class GoogleCloudBuild
that has the build fields that Orca cares about and have retrofit
deserialize the result from igor. Also, add the resulting field
to the stage context so it can be used by downstream tasks.

* feat(gcb): Monitor GCB build status after starting a build

Instead of immediatly returning success once the new GCB build is
accepted, wait until the build completes and set the status of the
stage based on the result of the build.