refactor(kubernetes): Simplify account creation code (#3630)
* test(kubernetes): Add tests to credentials initializing

There are currently no tests on initializing credentials; before
refactoring this code, add some tests.

Also, a number of the Kubernetes tests depend on being able to explicitly
pass credentials to a builder for KubernetesNamedAccountCredentials,
which will be removed. Instead just use a mock object and specify the
desired properties.

* refactor(kubernetes): Remove unused fields from account credentials

Most of the fields on KubernetesNamedAccountCredentials are just
copies of the fields on the contained KubernetesCredentials object.
There's no need to also have these on the parent class (and in
fact we never look these up on the parent class).  Remove all
of the redundant fields.

For the remaining fields:
- Apply 'private final' consistently
- Replace the accessors with a @Getter annotation

* refactor(kubernetes): Consolidate defaulting of account parameters

Defaulting of account paramters is scattered across a few different classes;
consolidate this to the definition of the properties class so it's easier
to see the values of these properties.

* refactor(kubernetes): Pass ManagedAccount to builder

Rather than pass each property of the ManagedAccount individually
to the Builder for KubernetesNamedAccountCredentials, just pass the
entire object.

* refactor(kubernetes): Use a factory class instead of a builder

Now that we're passing the entire ManagedAccount class to the builder
rather than the properties individually, there are no optional
parameters, so a builder class is less optimal. Other than the
ManagedAccount, all properties of the builder are beans that can
be autowired. Create a factory bean that autowires in these other
beans, and exposes methods to create a v1 or a v2 account from a
supplied ManagedAccount.

Ideally most of the logic in CredentialFactory would actually live
in as factories in the credential classes themselves; I'll likely
make this change in a future PR, but for now this is a reasonable
shim to avoid making this PR too large.

* fix(kubernetes): Add kube config to test

The V1 provider needs a valid kube config to properly initialize;
ideally we'd be able to inject mocks into the classes but all the
initializing code is using static classes that are called from Java
and can't be mocked. To fix the tests for now, just add a stub
kube config that has the minimum requirements to create the account.

* fix(kubernetes): Restore properties that are still needed

I removed a few properties from KubernetesNameAccountCredentials
that are actually needed by deck; while the accessors were never
called from clouddriver, these fields still included in the
serialized object. Restore the two missing fields.

I think we can simplify SpinnakerKindMap, but this PR is getting
large enough so just restoring what we had before.