fix(pipelines): Don't swallow exceptions when parsing pipeline templates (#546)
If the pipeline planning in Orca fails because of an error in the template, the error is just logged in Echo and the execution disappears. This makes it very hard to debug failing templates. I don't know if this happens to MPT v2 as well, but it is a problem for v1 at least.
This fix will let the pipeline triggering continue even if the planning fails, so that Orca later can register the failed execution.