chore(*): Bump spinnaker-dev plugin (#2914)
* chore(*): Apply codestyles to non-project files

The next commit pulls in the codestyles plugin, and applies it
to the base project (but not to subprojects). In preparation,
run the formatter and commit the changes as a single commit.

* chore(*): Bump spinnaker-dev plugin

We need the new plugin to set spring.config.additional-location instead
of spring.config.location when running services.